requirements.txt
asgiref==3.7.2
boto3==1.28.3
botocore==1.31.3
Django==4.2.3
django-cors-headers==4.2.0
django-jazzmin==2.6.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
jmespath==1.0.1
Pillow==10.0.0
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.16
